+++ core/specfem/CMakeLists.txt
@@ -4,3 +4,4 @@ add_subdirectory(receivers)
 add_subdirectory(shape_functions)
 add_subdirectory(source)
 add_subdirectory(jacobian)
+add_subdirectory(timescheme)
