+++ core/specfem/timescheme/newmark.cpp
@@ -1,3 +1,3 @@
-#include "timescheme/newmark.tpp"
+#include "specfem/timescheme/newmark.tpp"
 #include "specfem_setup.hpp"
 #include <ostream>
