+++ include/execute.hpp
@@ -8,9 +8,9 @@
 #include "solver/solver.hpp"
 #include "specfem/receivers.hpp"
 #include "specfem/source.hpp"
+#include "specfem/timescheme.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
-#include "timescheme/timescheme.hpp"
 #include "yaml-cpp/yaml.h"
 #include <Kokkos_Core.hpp>
 #include <boost/program_options.hpp>
