+++ include/parameter_parser/solver/solver.hpp
@@ -3,7 +3,7 @@
 #include "periodic_tasks/periodic_task.hpp"
 #include "solver/solver.hpp"
 #include "specfem/assembly.hpp"
-#include "timescheme/newmark.hpp"
+#include "specfem/timescheme/newmark.hpp"
 #include "utilities/strings.hpp"
 #include <memory>
 #include <string>
