+++ include/parameter_parser/time_scheme/time_scheme.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "specfem/timescheme/newmark.hpp"
 #include "specfem_setup.hpp"
-#include "timescheme/newmark.hpp"
 #include "yaml-cpp/yaml.h"
 #include <tuple>
