+++ include/solver/time_marching.hpp
@@ -7,8 +7,8 @@
 #include "kokkos_kernels/frechet_kernels.hpp"
 #include "periodic_tasks/periodic_task.hpp"
 #include "solver.hpp"
-#include "timescheme/newmark.hpp"
-#include "timescheme/timescheme.hpp"
+#include "specfem/timescheme.hpp"
+#include "specfem/timescheme/newmark.hpp"
 
 namespace specfem {
 namespace solver {
