+++ tests/unit-tests/seismogram/elastic/seismogram_tests.cpp
@@ -10,7 +10,7 @@
 #include "quadrature/interface.hpp"
 #include "solver/solver.hpp"
 #include "specfem/assembly.hpp"
-#include "timescheme/timescheme.hpp"
+#include "specfem/timescheme.hpp"
 #include "yaml-cpp/yaml.h"
 
 // ----- Parse test config ------------- //
